Clustering in HTML format
This is an HTML version of the Clustering practice. 
Used some of the clustering techniques like K-means clustering, , experiment with kernel transformation, used Gaussian Mix, PCA for visualization and elbow/silhouette as comparison methods. In the end the best results I got were from Gaussian Mix:  
 Accuracy ( true positives, %) of Gmm_labels:
Cluster 1 - 100.0% of 50 objects
Cluster 0 - 90.0% of 50
Cluster 2 - 100.0% of 50