Real Estate Analysis and Modeling
Analysis and Modeling of the Real Estate prices. 
Has a lot of data visualization and exploration. After the exploration I experiment with 5 different hypothesis: 4  - with  linear regression, the last one - KNN.
The other difference between all of them is the pre-processing methods used.
The aim of the project was not to get the best prediction with as little code lines as possible, but to experiment and explore data exploration and pre-transformation techniques.